+++ internal/types/keysign.go
@@ -14,7 +14,6 @@ import (
 type KeysignRequest struct {
 	PublicKeyECDSA   string   `json:"public_key_ecdsa"`   // ECDSA public key, used to identify the backup file
 	Messages         []string `json:"messages"`           // Messages need to be signed
-	Key              string   `json:"key"`                // should always be Vultisigner
 	Session          string   `json:"session"`            // Session ID , it should be an UUID
 	HexEncryptionKey string   `json:"hex_encryption_key"` // Hex encryption key, used to encrypt the keysign messages
 	DerivePath       string   `json:"derive_path"`        // Derive Path
@@ -28,9 +27,6 @@ func (r KeysignRequest) IsValid() error {
 	}
 	if len(r.Messages) == 0 {
 		return errors.New("invalid messages")
-	}
-	if r.Key == "" {
-		return errors.New("invalid key")
 	}
 	if r.Session == "" {
 		return errors.New("invalid session")
